Add files via upload
This Python script is a text-based game where players enter weights and values for five items. The programme randomly picks items to fill a "snapsack" with a weight limit of 10. If the total value of the chosen items hits 90 or more without going over the weight limit, the player wins. It has a loop that lets the user play the game again.